refactor: change Percent-Charged header from Double to Int
- Update TrmnlApiService to accept Int? instead of Double?
- Update TrmnlDeviceUpdateRequest percentCharged field to Int?
- Remove .toDouble() conversions in TrmnlDisplayRepository
- Update test assertions to use Int values
- Eliminates unnecessary type conversions since battery level is already an Int

+++ app/src/main/java/ink/trmnl/android/network/TrmnlApiService.kt
@@ -105,7 +105,7 @@ interface TrmnlApiService {
         @Header("ID") deviceMacId: String? = null,
         @Header("BASE64") useBase64: Boolean? = null,
         @Header("RSSI") rssi: Int? = null,
-        @Header("Percent-Charged") percentCharged: Double? = null,
+        @Header("Percent-Charged") percentCharged: Int? = null,
     ): ApiResult<TrmnlDisplayResponse, Unit>
 
     /**
